% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{asymmetric_cost_explorer_cross_validated}
\alias{asymmetric_cost_explorer_cross_validated}
\title{Asymmetric Cost Explorer}
\usage{
asymmetric_cost_explorer_cross_validated(phat, ybin, K_CV = 5, ...)
}
\arguments{
\item{phat}{The vector of probability estimates to be thresholded to make a binary decision}

\item{ybin}{The true binary responses}

\item{K_CV}{We wish to fit the \code{phat} thresholds out of sample using this number of folds. Default is \code{5}.}

\item{...}{Other parameters to be passed into the \code{asymmetric_cost_explorer} function}
}
\value{
A table with column 1: \code{proportions_desired}, column 2: actual proportions (as close as possible), column 3: error rate, column 4: probability threshold.
}
\description{
Given a set of desired proportions of predicted outcomes, what is the error rate for each of those models?
}
\author{
Adam Kapelner
}
