% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{eigen_det}
\alias{eigen_det}
\title{A fast det(X) function}
\usage{
eigen_det(X, num_cores = 1)
}
\arguments{
\item{X}{A numeric matrix of size p x p}

\item{num_cores}{The number of cores to use. Unless p is large, keep to the default of 1.}
}
\value{
The determinant as a scalar numeric value
}
\description{
Via the eigen package
}
\examples{
  p = 30
  eigen_det(matrix(rnorm(p^2), nrow = p))
}
