% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{eigen_inv}
\alias{eigen_inv}
\title{A fast solve(X) function}
\usage{
eigen_inv(X, num_cores = 1)
}
\arguments{
\item{X}{A numeric matrix of size p x p}

\item{num_cores}{The number of cores to use. Unless p is large, keep to the default of 1.}
}
\value{
The resulting matrix
}
\description{
Via the eigen package
}
\examples{
  p = 10
  eigen_inv(matrix(rnorm(p^2), nrow = p))
}
