% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{eigen_Xt_times_diag_w_times_X}
\alias{eigen_Xt_times_diag_w_times_X}
\title{A fast Xt [times] diag(w) [times] X function}
\usage{
eigen_Xt_times_diag_w_times_X(X, w, num_cores = 1)
}
\arguments{
\item{X}{A numeric matrix of size n x p}

\item{w}{A numeric vector of length p}

\item{num_cores}{The number of cores to use. Unless p is large, keep to the default of 1.}
}
\value{
The resulting matrix
}
\description{
Via the eigen package
}
\examples{
  n = 100
  p = 10
  X = matrix(rnorm(n * p), nrow = n, ncol = p)
  w = rnorm(p)
  eigen_Xt_times_diag_w_times_X(t(X), w)
}
