% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{fast_logistic_regression}
\alias{fast_logistic_regression}
\title{FastLR Wrapper}
\usage{
fast_logistic_regression(
  Xmm,
  ybin,
  drop_collinear_variables = FALSE,
  lm_fit_tol = 1e-07,
  solve_tol = .Machine$double.eps,
  ...
)
}
\arguments{
\item{Xmm}{The model.matrix for X (you need to create this yourself before)}

\item{ybin}{The binary response vector}

\item{drop_collinear_variables}{Should we drop perfectly collinear variables? Default is \code{FALSE} to inform the user of the problem.}

\item{lm_fit_tol}{When \code{drop_collinear_variables = TRUE}, this is the tolerance to detect collinearity among predictors.
We use the default value from \code{base::lm.fit}'s which is 1e-7. If you fit the logistic regression and
still get p-values near 1 indicating high collinearity, we recommend making this value smaller.}

\item{solve_tol}{Tolerance when inverting X^T W X, a quantity needed when computing standard errors of the coefficients. 
The default is \code{.Machine$double.eps} which is used in \code{base::solve}. Make this value smaller if you 
still get errors even when setting \code{drop_collinear_variables = TRUE}. You may have to play around with 
this parameter and \code{lm_fit_tol} a bit to get your desired result.}

\item{...}{Other arguments to be passed to \code{fastLR} (see documentation there)}
}
\value{
A list of raw results
}
\description{
Returns most of what you get from glm
}
\examples{
 \dontrun{
flr = fast_logistic_regression(
	Xmm = model.matrix(~ . - type, MASS::Pima.te), 
 ybin = as.numeric(MASS::Pima.te$type == "Yes")
)
	}
}
