% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{fast_logistic_regression_stepwise_forward}
\alias{fast_logistic_regression_stepwise_forward}
\title{Rapid Forward Stepwise Logistic Regression}
\usage{
fast_logistic_regression_stepwise_forward(
  Xmm,
  ybin,
  pval_threshold = 0.05,
  use_intercept = TRUE,
  verbose = TRUE,
  drop_collinear_variables = FALSE,
  lm_fit_tol = 1e-07,
  solve_tol = .Machine$double.eps,
  ...
)
}
\arguments{
\item{Xmm}{The model.matrix for X (you need to create this yourself before)}

\item{ybin}{The binary response vector}

\item{pval_threshold}{The significance threshold to include a new variable. Default is \code{0.05}.}

\item{use_intercept}{Should we automatically begin with an intercept? Default is \code{TRUE}.}

\item{verbose}{Print out messages during the loop? Default is \code{TRUE}.}

\item{drop_collinear_variables}{Parameter used in \code{fast_logistic_regression}. See documentation there.}

\item{lm_fit_tol}{Parameter used in \code{fast_logistic_regression}. See documentation there.}

\item{solve_tol}{Parameter used in \code{fast_logistic_regression}. See documentation there.}

\item{...}{Other arguments to be passed to \code{fastLR} (see documentation there)}
}
\value{
A list of raw results
}
\description{
Returns most of what you get from glm
}
\examples{
 \dontrun{
flr = fast_logistic_regression_stepwise_forward(
  Xmm = model.matrix(~ . - type, MASS::Pima.te), 
  ybin = as.numeric(MASS::Pima.te$type == "Yes")
)
	}
}
