% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastLogisticRegressionWrap.R
\docType{package}
\name{fastLogisticRegressionWrap}
\alias{fastLogisticRegressionWrap}
\title{A Wrapper for FastLR}
\description{
A tool to find many types of a priori experimental designs
}
\references{
Kapelner, A
}
\author{
Adam Kapelner \email{kapelner@qc.cuny.edu}
}
\keyword{logistic}
\keyword{regression}
