% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{print.fast_logistic_regression}
\alias{print.fast_logistic_regression}
\title{FastLR Wrapper Print}
\usage{
\method{print}{fast_logistic_regression}(x, ...)
}
\arguments{
\item{x}{The object built using the \code{fast_logistic_regression} or \code{fast_logistic_regression_stepwise} wrapper functions}

\item{...}{Other arguments to be passed to print}
}
\value{
The summary as a data.frame
}
\description{
Returns the summary table a la glm
}
\examples{
 \dontrun{
flr = fast_logistic_regression(
	Xmm = model.matrix(~ . - type, MASS::Pima.te), 
 ybin = as.numeric(MASS::Pima.te$type == "Yes"))
print(flr)
	}
}
