% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{summary.fast_logistic_regression_stepwise}
\alias{summary.fast_logistic_regression_stepwise}
\title{FastLR Wrapper Summary}
\usage{
\method{summary}{fast_logistic_regression_stepwise}(object, ...)
}
\arguments{
\item{object}{The object built using the \code{fast_logistic_regression_stepwise} wrapper functions}

\item{...}{Other arguments to be passed to \code{summary}.}
}
\value{
The summary as a data.frame
}
\description{
Returns the summary table a la glm
}
\examples{
library(MASS); data(Pima.te)
flr = fast_logistic_regression_stepwise_forward(
	Xmm = model.matrix(~ . - type, Pima.te), 
 ybin = as.numeric(Pima.te$type == "Yes"))
summary(flr)
}
