% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{summary.fast_logistic_regression}
\alias{summary.fast_logistic_regression}
\title{FastLR Wrapper Summary}
\usage{
\method{summary}{fast_logistic_regression}(object, ...)
}
\arguments{
\item{object}{The object built using the \code{fast_logistic_regression} or \code{fast_logistic_regression_stepwise} wrapper functions}

\item{...}{Other arguments to be passed to \code{summary}.}
}
\value{
The summary as a data.frame
}
\description{
Returns the summary table a la glm
}
\examples{
 \dontrun{
flr = fast_logistic_regression(
	Xmm = model.matrix(~ . - type, MASS::Pima.te), 
 ybin = as.numeric(MASS::Pima.te$type == "Yes"))
summary(flr)
	}
}
